entity/requirements.txt
python-docx>=0.8.11
pdfplumber>=0.9.0
jieba>=0.42.1
pandas>=1.5.0
openpyxl>=3.0.10
dashscope>=1.14.0
flask>=3.0.0
openai>=1.0.0
